requirements.txt
streamlit
pandas
io

matplotlib.pyplot
seaborn
urlextract
wordcloud
pandas
collections
emoji
re